MAINT: Change shape_is_bordering and _containing to be floats
Instead of uints. This means we can encode fractional values for
bordering based on how close the shapes are. Also, the initial
values are now -1.0, which distinguishes them from the set value
of 0.

+++ svrt_generator.cc
@@ -199,8 +199,8 @@ struct VignetteSet {
   int nb_symbolic_outputs;
   unsigned char *nb_shapes_each;
   float *shapes_symb_output;
-  uint *shape_is_bordering;
-  uint *shape_is_containing;
+  float *shape_is_bordering;
+  float *shape_is_containing;
 };
 
 void svrt_generate_vignettes(int n_problem, int nb_vignettes, long *labels,
@@ -222,14 +222,14 @@ void svrt_generate_vignettes(int n_problem, int nb_vignettes, long *labels,
   result->nb_symbolic_outputs = Vignette::nb_symbolic_outputs;
   result->nb_shapes_each = (unsigned char *) malloc(sizeof(unsigned char) * result->nb_vignettes);
   result->shapes_symb_output = (float *) malloc(sizeof(float) * result->nb_vignettes * result->max_shapes * result->nb_symbolic_outputs);
-  result->shape_is_bordering = (uint *) malloc(sizeof(uint) * result->nb_vignettes * result->max_shapes * result->max_shapes);
-  result->shape_is_containing = (uint *) malloc(sizeof(uint) * result->nb_vignettes * result->max_shapes * result->max_shapes);
+  result->shape_is_bordering = (float *) malloc(sizeof(float) * result->nb_vignettes * result->max_shapes * result->max_shapes);
+  result->shape_is_containing = (float *) malloc(sizeof(float) * result->nb_vignettes * result->max_shapes * result->max_shapes);
 
   unsigned char *s = result->data;
   unsigned char *out_pointer_nb_shapes = result->nb_shapes_each;
   float *out_pointer_shapes_symb_output = result->shapes_symb_output;
-  uint *out_pointer_shape_is_bordering = result->shape_is_bordering;
-  uint *out_pointer_shape_is_containing = result->shape_is_containing;
+  float *out_pointer_shape_is_bordering = result->shape_is_bordering;
+  float *out_pointer_shape_is_containing = result->shape_is_containing;
   for(int i = 0; i < nb_vignettes; i++) {
     if(labels[i] == 0 || labels[i] == 1) {
       vg->generate(labels[i], &tmp);
@@ -250,8 +250,8 @@ void svrt_generate_vignettes(int n_problem, int nb_vignettes, long *labels,
       *out_pointer_shapes_symb_output++ = in_pointer_shapes_symb_output[k];
     }
 
-    uint *in_pointer_shape_is_bordering = tmp.shape_is_bordering;
-    uint *in_pointer_shape_is_containing = tmp.shape_is_containing;
+    float *in_pointer_shape_is_bordering = tmp.shape_is_bordering;
+    float *in_pointer_shape_is_containing = tmp.shape_is_containing;
     for(int k = 0; k < result->max_shapes * result->max_shapes; k++) {
       *out_pointer_shape_is_bordering++ = in_pointer_shape_is_bordering[k];
       *out_pointer_shape_is_containing++ = in_pointer_shape_is_containing[k];
